+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixes
+
+- fix(remix): Use `npx --package=@sentry/remix` for `sentry-upload-sourcemaps` to avoid global bin namespace collisions
+
 ### Features
 
 - feat(react-router): Use `sentryOnError` on `HydratedRouter` instead of mutating `root.tsx` ErrorBoundary